backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.8.2

# LangChain split packages
langchain==0.2.6
langchain-community==0.2.6
langchain-openai==0.1.7
langchain-chroma==0.1.1

# Vector DB
chromadb==0.4.24   # (stable, avoids the tenant issue)
tiktoken==0.7.0
pypdf
